requirements.txt
apache-airflow-providers-postgres==1.0.1
pypika==0.47.6
html-sanitizer==1.9.1
nltk==3.6.2
Pillow==8.3.1
algoliasearch==2.5.0
vcrpy==4.1.1
pytest==6.2.4
pytest-mock==3.6.1
pytest-vcr==1.0.2
contentful==1.13.1
markdown==3.3.6
redis==4.1.0